Improve text rendering
Use the DotGothic16-Regular font for sizes 18 and below. This font looks
much better at small point sizes without anti-aliasing. For larger
sizes use Kosugi-Regular.

The y-offset is calculated differently now, with the font being centered
based on the height of the letter 'A'. This fixes an issue with the old
method when the font's ascent is much larger than the height of a
typical character (as is the case with DotGothic).