Implement Util.function[10] - down/left fade
This function is an animated fade from left to right or from top to
bottom. It can either fade to color index 0, or to another surface
(hardcoded as surface 2).

The fade itself is implemented using a specific pattern of on/off
pixels. This implementation should be a pixel-perfect reproduction of
the original AI5Win implementation.