Implement Util 5 - scale height
Implement function to scale the height of the screen surface.

Because SDL_BlitScaled doesn't support blitting an indexed surface to an
RGB surface, we have to use an intermediary surface when scaling is
applied.

As far as I can tell, this function is only used to implement a screen
shaking effect.