Implement System.Palette.set with argument
If System.Palette.set receives an argument, that argument is a color
that the palette should be filled with. This fixes various bugs where
incorrect palettes would be loaded.

Also, gfx_dirty now takes an surface index argument, and only marks the
screen as dirty if the index matches the current screen index.