Allow loading files from filesystem
Allow files to be loaded from the filesystem, if an archive was not
provided.

The YU-NO English translation uses this for sound effects and voices.
Additionally, the YU-NO TL loads voices from .ogg files, even when the
code asks for a .wav file. So there is a hack to make this work.