Implement reflector animation for YU-NO
Implement the reflector animation (bottom right corner of screen). This
animation is hardcoded into the executable, and activates when flag 2 is
set.

This implementation is not pixel-perfect, but it's pretty close. It
works by loading the base frame from a CG, and then procedurally
generating the subsequent frames (the CG archive does not contain all
frames--presumably they are baked into the executable).