Use signed array indices
When opening the palmtop in YU-NO, code like this is run:

    if (i > 0 && array[i - 1] != v) {
        ...
    }

Which is fine, except that '&&' is not short-circuiting in mes bytecode!
If the index into the array is treated as unsigned, this will read from
invalid memory.

To fix this issue, we interpret array indices as signed values.
Accessing memory through a negative array index is probably a bug in
most cases, but it should allow code like the above to run without
segfaulting.