Implement YU-NO dream sequence
In the Japanese version, this is part of the SP disk contents, but the
English version inserts it into the main game (so it's required to
complete the game).

This implementation does not display the psychedelic background effect
or text box frame, it just fades the text in/out similar to the text
which precedes it.

The text itself is embedded into the executable. To avoid a potential
copyright issue, we only store the offsets into the .exe file and read
the text at runtime.