aishimai: implement backlog
Implement the backlog feature for Ai Shimai. The way this works is a bit
convoluted.

The flag FLAG_LOG_ENABLE controls whether the backlog is enabled.
The flag FLAG_LOG_TEXT controls whether TXT/STR statements are written
to the backlog.

If both flags are on, then another flag, FLAG_LOG, is turned on by the
VM when it encounters a TXT or STR statement. This flag causes every
byte read from the current mes file to be written to the current backlog
entry. FLAG_LOG is then turned off when the statement is finished
executing.

In addition to this, System.function[19] (System.Backlog) contains a
number of functions for manipulating and querying the backlog.