aishimai: implement screen shake effect
Implement Util(0, x, y). This util shifts the screen by the given
offset, and is used to implement a screen shake effect.

Because the game simply calls this function repeatedly, we throttle it.

Interestingly, the method used to implement this function in the
original engine appears to be broken on modern versions of windows. Only
positive x offsets are applied.