Refactor audio and System/Util functions
Refactor the audio code to reduce and consolidate boilerplate between
the two inplementations. Audio channels are now represented by an enum
and passed into the audio functions.

Much of the code in sys.c is moved to classics.c.

All of the code in util.c is moved to yuno.c. Utils are always
game-specific and this file just contained the utils for YUNO.