doukyuusei: movie support
Support movie playback with ffmpeg.

There are two movies in Doukyuusei, both at the end.

The first is a simple video, with no audio or special sequencing
(sakura.avi).

The second is the ending credits roll, which has separate video and
audio files, with the video file being looped at specific points
throughout the movie. Additionally, there is a character sprite
animation and the credits text which are overlayed on top of the video.

This ending movie was implemented by painstakingly analyzing a recording
of the ending from AI5WIN.EXE. It should be *close* to frame-perfect,
but there could be minor differences.