Improve texthooking
Text is now buffered so that, for example, lines are not broken up when
the player's name is displayed in the middle of a line. When System.wait
or menuexec is called, the buffered text is commited and output.

When calling the menu initialization function, buffering is temporarily
disabled so that each menu item appears as its own line.

When the text location changes, we also commit the current text. This
prevents text like "見る・調べる" in Doukyuusei from cluttering up
unrelated lines.