vm: use dispatch tables for stmt/expr ops
`struct game` now contains dispatch tables for statement and expression
opcodes. These are used to eliminate all of the hacks which used boolean
variables to activate game-specific quirks.

One inconsistency with this is the STR/TXT opcodes. Because the same
functionality can be triggered without an opcode, and because there is a
lot of boilerplate in these functions, we allow for the `vm_draw_text`
function to be overriden instead of STR/TXT.