Fix issue with System.font_weight
Changes to System.font_weight take effect immediately.
System.set_font_size does not need to be called afterwards.

This fixes an issue in Isaku where the main menu was drawn with the
wrong font weight.

Incidentally, in all supported games so far, the font_weight is ALWAYS
set to 1 before drawing text, and is only 0 briefly at startup or after
loading a save file.