Kakyuusei support
Implement support for Kakyuusei (1998 CD version).

This is the first 256-color indexed game to be supported. There's a lot
of tricky palette management going on, with `gfx.palette`,
`memory.palette`, and a couple extra palette banks being used.

The backlog system in this game works differently than in other games.
Some new TXT/STR instruction implementations are added to support it,
and unprefixed text is now handled through virtual functions as well.

The map system is largely identical to Doukyuusei's, but with various
quirks that are handled via game-specific code paths.