Initial Shuusaku support
This is the first "AIWIN"-engine game to be supported. Since we already
use opcode tables and a virtualized memory layout, the changes to the VM
are relatively minimal.

For now, many AIW statement opcodes are implemented in shuusaku.c rather
than vm.c. It remains to be seen which of these can have a generic
implementation. Probably some can, others cannot.

This game has a lot of functionality implemented in native code. This
has been split across several files in the shuusaku/ directory.

Both the DVD and the CD version should work, but only the DVD version
has been thoroughly tested.

Unimplemented (for now):
* Scroll bar on the schedule window (scroll wheel/arrow buttons can be
  used instead)
* Menu bar (middle click can be used instead to bring up a popup menu)
* Config UI (options can be modified via .ini file)